\name{gihyo-spark-book-example-package}
\alias{gihyo-spark-book-example-package}
\alias{gihyo-spark-book-example}
\docType{package}
\title{
\packageTitle{gihyo-spark-book-example}
}
\description{
\packageDescription{gihyo-spark-book-example}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gihyo-spark-book-example}
\packageIndices{gihyo-spark-book-example}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gihyo-spark-book-example}

Maintainer: \packageMaintainer{gihyo-spark-book-example}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
